feat: setup basic _mostly_ frontend based `multi stepped form` for taking user input
I was finally able to figure out how to properly do it, without much
load on the application, my first approach was using cookies but it was
most definitely not scalable as a single user creating a single test
would end up reading and writing a cookie about 4-5 times.
So, now I decided that everything would be vanilla javascript based(I am
not even using HTMX that here, just pure vanilla javascript), and the
whole process would be on the frontend.
> currently, it is loosely built and it also gives errors sometimes, and
I havn't even properly setup the javascript yet. It does require some
refinements, but it is the basic idea