feat: handle extraction of selected chapters, test format input
finally I was able to work on this again, it feels like an eternity.
I noticed a small problem with the way the inputs and `_base.json` is
structured, it will deliver poor performance as the size of `_base.json`
increases i.e more classes, subjects, and chapters get added to the
dataset.
There has to be some improvement in the actual data strcture of the
`_base.json` which reduces the number of for loops required to get the
required data by the client side javascript.
IDEA: maybe I can try splitting the input into seprate js variables for
classes, subjects and chapters ???